[Refactor] Duration and Repeat Duration columns are now int (0–10000)
Changes across grid + DV + pydantic contract:
- Grid cell editors: QDoubleSpinBox → QSpinBox (Run Time stays float).
- step_defaults / group_defaults: "1.0" → "1" for Duration and Repeat
  Duration.
- _validate_numeric_field formats Duration/Repeat Duration as int.
- All float(...) casts on these columns become int(...); path
  execution service uses int(float(...)) for legacy-"1.0" tolerance.
- extract_execution_params uses an int_cast that tolerates legacy
  float strings.
- StepParamsCommitMessage.duration / .repeat_duration: float → int.
- RouteLayerManager.duration Range: float → int (repeat_duration
  already int).
- _current_params returns ints; payload formatters use str(...) not
  ":g".
- Tests updated to use ints; added a legacy-float-string tolerance
  test on the helper.

+++ device_viewer/models/route.py
@@ -213,7 +213,7 @@ class RouteLayerManager(HasTraits):
     phase_label = Str("")
 
     # path execution properties (mirror protocol grid step defaults)
-    duration = Range(low=0.0, high=10000.0, value=1.0)
+    duration = Range(low=0, high=10000, value=1)
     trail_length = Range(low=1, high=10000, value=1)
 
     # Dynamic upper bound for trail_overlay: overlay cannot exceed trail_length - 1
@@ -254,9 +254,9 @@ class RouteLayerManager(HasTraits):
 
     def _current_params(self) -> dict:
         return {
-            "duration": float(self.duration),
+            "duration": int(self.duration),
             "repetitions": int(self.repetitions),
-            "repeat_duration": float(self.repeat_duration),
+            "repeat_duration": int(self.repeat_duration),
             "trail_length": int(self.trail_length),
             "trail_overlay": int(self.trail_overlay),
             "soft_start": bool(self.soft_start),
